organized file locations, removed unecessary imports
