+++ src/app/page.tsx
@@ -1,10 +1,12 @@
 import Image from "next/image";
-import Landing from "./pages/Landing";
+import Landing from "../sections/Landing"
+import Events from "../sections/Events"
 
 export default function Home() {
   return (
     <main>
       <Landing />
+      <Events />
     </main>
   );
 }
